requirements.txt
pillow
PySide2